requirements.txt
-i https://pypi.org/simple
accelerate==0.20.3
blinker==1.6.2 ; python_version >= '3.7'
cachetools==5.3.1 ; python_version >= '3.7'
certifi==2023.5.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
cloudevents==1.9.0
deprecation==2.1.0
fasttext==0.9.2
filelock==3.12.2 ; python_version >= '3.7'
flask==2.3.2
fsspec==2023.6.0 ; python_version >= '3.8'
functions-framework==3.4.0
google-api-core==2.11.1 ; python_version >= '3.7'
google-auth==2.21.0 ; python_version >= '3.6'
google-cloud-core==2.3.2 ; python_version >= '3.7'
google-cloud-storage==2.10.0
google-crc32c==1.5.0 ; python_version >= '3.7'
google-resumable-media==2.5.0 ; python_version >= '3.7'
googleapis-common-protos==1.59.1 ; python_version >= '3.7'
gunicorn==20.1.0 ; platform_system != 'Windows'
huggingface-hub==0.15.1 ; python_full_version >= '3.7.0'
idna==3.4 ; python_version >= '3.5'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
markupsafe==2.1.3 ; python_version >= '3.7'
mpmath==1.3.0
networkx==3.1 ; python_version >= '3.8'
numpy==1.25.0 ; python_version >= '3.9'
packaging==23.1 ; python_version >= '3.7'
protobuf==4.23.3 ; python_version >= '3.7'
psutil==5.9.5 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyasn1==0.5.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pybind11==2.10.4 ; python_version >= '3.6'
pyyaml==6.0 ; python_version >= '3.6'
regex==2023.6.3 ; python_version >= '3.6'
requests==2.31.0 ; python_version >= '3.7'
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
safetensors==0.3.1
setuptools==68.0.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sympy==1.12 ; python_version >= '3.8'
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0 ; python_version >= '3.7'
transformers[torch]==4.30.2
typing-extensions==4.6.3 ; python_version >= '3.7'
urllib3==1.26.16 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
watchdog==3.0.0 ; python_version >= '3.7'
werkzeug==2.3.6 ; python_version >= '3.8'
